form answers as single DB column, reorg helpers
