+++ src/api/helpers.ts
@@ -1,5 +1,11 @@
 import { FormFields } from "../Components/Pages/Form/Form";
-import { GCEPostData } from "../types/APIDataTypes";
+import { Determination, EligibilityResults } from "../hooks/eligibility";
+import {
+  Coverage,
+  CriteriaDetermination,
+  FormAnswers,
+  GCEPostData,
+} from "../types/APIDataTypes";
 import { HTTPError, NetworkError } from "./error-reporting";
 
 export const WowApiFetcher = async (url: string) => {
@@ -69,18 +75,35 @@ export const cleanFormFields = ({
   landlord,
   rentStabilized,
   housingType,
-}: FormFields): GCEPostData => {
-  const form_bedrooms = bedrooms === null ? undefined : bedrooms;
-  const form_rent = rent === null ? undefined : parseFloat(rent);
-  const form_owner_occupied = landlord === null ? undefined : landlord;
-  const form_rent_stab = rentStabilized === null ? undefined : rentStabilized;
-  const form_subsidy = housingType === null ? undefined : housingType;
-
+}: FormFields): FormAnswers => {
   return {
-    form_bedrooms,
-    form_rent,
-    form_owner_occupied,
-    form_rent_stab,
-    form_subsidy,
+    bedrooms: bedrooms === null ? undefined : bedrooms,
+    rent: rent === null ? undefined : parseFloat(rent),
+    owner_occupied: landlord === null ? undefined : landlord,
+    rent_stab: rentStabilized === null ? undefined : rentStabilized,
+    subsidy: housingType === null ? undefined : housingType,
   };
+};
+
+// Recodes values for database model
+export const determinationToCoverage = (
+  determination: Determination
+): Coverage => {
+  const COVERAGE: { [key in Determination]: Coverage } = {
+    eligible: "COVERED",
+    ineligible: "NOT_COVERED",
+    unknown: "UNKNOWN",
+  };
+  return COVERAGE[determination];
+};
+
+// Restructures criteria eligibility for database model
+export const extractDeterminations = (
+  eligibilityResults: EligibilityResults
+): CriteriaDetermination => {
+  const criteriaDeterminations: CriteriaDetermination = {};
+  Object.values(eligibilityResults).forEach((x) => {
+    criteriaDeterminations[x.criteria] = x.determination!;
+  });
+  return criteriaDeterminations;
 };
